+++ Sources/Schedulers/DPMSolverMultistepScheduler.swift
@@ -23,15 +23,15 @@ import RandomGenerator
 public final class DPMSolverMultistepScheduler: Scheduler {
     public let trainStepCount: Int
     public let inferenceStepCount: Int
-    public let betas: [Float]
-    public let alphas: [Float]
-    public let alphasCumProd: [Float]
+    public let betas: [Double]
+    public let alphas: [Double]
+    public let alphasCumProd: [Double]
     public let timeSteps: [Double]
     public let predictionType: PredictionType
 
-    public let alpha_t: [Float]
-    public let sigma_t: [Float]
-    public let lambda_t: [Float]
+    public let alpha_t: [Double]
+    public let sigma_t: [Double]
+    public let lambda_t: [Double]
     
     public let solverOrder = 2
     private(set) var lowerOrderStepped = 0
@@ -57,8 +57,8 @@ public final class DPMSolverMultistepScheduler: Scheduler {
         stepCount: Int = 50,
         trainStepCount: Int = 1000,
         betaSchedule: BetaSchedule = .scaledLinear,
-        betaStart: Float = 0.00085,
-        betaEnd: Float = 0.012,
+        betaStart: Double = 0.00085,
+        betaEnd: Double = 0.012,
         predictionType: PredictionType = .epsilon
     ) {
         self.trainStepCount = trainStepCount
@@ -75,7 +75,7 @@ public final class DPMSolverMultistepScheduler: Scheduler {
 
         // Currently we only support VP-type noise shedule
         self.alpha_t = vForce.sqrt(self.alphasCumProd)
-        self.sigma_t = vForce.sqrt(vDSP.subtract([Float](repeating: 1, count: self.alphasCumProd.count), self.alphasCumProd))
+        self.sigma_t = vForce.sqrt(vDSP.subtract([Double](repeating: 1, count: self.alphasCumProd.count), self.alphasCumProd))
         self.lambda_t = zip(self.alpha_t, self.sigma_t).map { α, σ in log(α) - log(σ) }
 
         var timeSteps: [Int] = linspace(0, Float(trainStepCount - 1), stepCount+1).dropFirst().map { Int(round($0)) }
@@ -92,7 +92,7 @@ public final class DPMSolverMultistepScheduler: Scheduler {
     func convertModelOutput(modelOutput: MLShapedArray<Float32>, timestep: Int, sample: MLShapedArray<Float32>) -> MLShapedArray<Float32> {
         assert(modelOutput.scalarCount == sample.scalarCount)
         let scalarCount = modelOutput.scalarCount
-        let (alpha_t, sigma_t) = (self.alpha_t[timestep], self.sigma_t[timestep])
+        let (alpha_t, sigma_t) = (Float32(self.alpha_t[timestep]), Float32(self.sigma_t[timestep]))
         // This could be optimized with a Metal kernel if we find we need to
         switch predictionType {
         case .epsilon:
